[ICU-12701] fix missing implicit casts
These are needed for IBM i.
X-Commit-URL: https://ssl.icu-project.org/trac/changeset/39113
Former-commit-id: 9090a2ad7981fc57b93a0fa6ff6b7398e80744b8

+++ icu4c/source/i18n/decimfmtimpl.cpp
@@ -243,9 +243,9 @@ DecimalFormatImpl::setMultiplierScale(int32_t scale) {
     if (scale == 0) {
         // Needed to preserve equality. fMultiplier == 0 means
         // multiplier is 1.
-        fMultiplier.set(0);
+        fMultiplier.set((int32_t)0);
     } else {
-        fMultiplier.set(1);
+        fMultiplier.set((int32_t)1);
         fMultiplier.shiftDecimalRight(scale);
     }
 }
